feat(Song)!: use `yt-dlp` instead of `youtube-dl`
Even though `youtube-dl` is still somehow actively developed, its last
release is on December 2021. It is not only just old but also fails on
some task because of changed YouTube API.

The last things that resulted this change is it was crashing randomly
even while doing basic search.

Also its `songbird` side was already changed and only the code in `Song`
struct uses `youtube-dl` now.

With this change all `youtube-dl` dependency will be removed and only
`yt-dlp` will be used.