refactor: create a `get_common!()` macro for commands
At the start of every command -for now only in music commands- there is a part
that is gets the server and guid. This logic is the same for every command; so,
making it seperate macro removes a lot of copy paste now and in the feature.

To be able to achive this `Server`s in server `HashMap` wrapped in `Arc`. This
may also have other benefits such as not holding server `HashMap`s lock for
entire command run.