chore: add new clippy lints
The most noticable change in lints that adding new `restriction` type
lints. Every new version add some of them.

The last checked clippy version added to reduce future work.

The lines ends with `// forbid` will be in `forbid` level in the future
but cannot be in that kevek because clippy still errors about external
macros.

The applications of the lints will be in future MRs.