feat: add custom youtube scraper for search
The old implementation calls `youtube-dl` (or its variations) for search,
but `youtube-dl` is very slow. I have also tested `yt-dlp`.

It was taking so long that some users were thinking that boot is not
responding.

With this commit it sends a GET request to YouTube as if it was
searching using a web browser and scrapes the data out of it. Even
though it seem to do a lot of things like finding the relative code,
parsing it using `serde` crate, and then creating a song list out of it;
it is surprisingly faster than `youtube-dl`.

It also work until YouTube changes its inner API.